bg_remover_app/requirements.txt
PyQt6
Pillow
rembg[cpu]
